REFACTOR(breadcrumbs): `breadcrumbs.py` moved to `utils` folder
